refactor: rename namespaces for better organization
- Change gui::Config to parser::Config for args parsing
- Change Network::ServerCommunication to network::ServerCommunication
- Update all references and include statements accordingly
- Fix typo in ArgsParser.cpp include (Argsparser.hpp)
- Update help message to reference zappy_parser instead of zappy_gui

This change improves code organization by placing classes in more
semantically appropriate namespaces and follows consistent naming
conventions.
