+++ gui/src/server/ServerCommunication.cpp
@@ -10,13 +10,13 @@
 #include <stdexcept>
 #include <string>
 
-Network::ServerCommunication::ServerCommunication(int port,
+network::ServerCommunication::ServerCommunication(int port,
                                                   const std::string &hostname)
     : _clientFd(-1), _port(port), _hostname(hostname), _connected(false) {
   std::memset(&_serverAddr, 0, sizeof(_serverAddr));
 }
 
-Network::ServerCommunication::~ServerCommunication() {
+network::ServerCommunication::~ServerCommunication() {
   if (_clientFd != -1) {
     close(_clientFd);
     _clientFd = -1;
@@ -24,11 +24,11 @@ Network::ServerCommunication::~ServerCommunication() {
   _connected = false;
 }
 
-bool Network::ServerCommunication::isConnected() const noexcept {
+bool network::ServerCommunication::isConnected() const noexcept {
   return _connected;
 }
 
-bool Network::ServerCommunication::connectToServer() {
+bool network::ServerCommunication::connectToServer() {
   try {
     createSocket();
     establishConnection();
@@ -40,14 +40,14 @@ bool Network::ServerCommunication::connectToServer() {
   }
 }
 
-void Network::ServerCommunication::createSocket() {
+void network::ServerCommunication::createSocket() {
   _clientFd = socket(AF_INET, SOCK_STREAM, 0);
   if (_clientFd == -1)
     throw std::runtime_error("Failed to create socket: " +
                              std::string(strerror(errno)));
 }
 
-void Network::ServerCommunication::establishConnection() {
+void network::ServerCommunication::establishConnection() {
   _serverAddr.sin_family = AF_INET;
   _serverAddr.sin_port = htons(_port);
   if (inet_pton(AF_INET, _hostname.c_str(), &_serverAddr.sin_addr) <= 0)
@@ -61,7 +61,7 @@ void Network::ServerCommunication::establishConnection() {
   _connected = true;
 }
 
-void Network::ServerCommunication::sendMessage(const std::string &message) {
+void network::ServerCommunication::sendMessage(const std::string &message) {
   if (!_connected || _clientFd == -1)
     throw std::runtime_error("Not connected to server");
 
@@ -82,7 +82,7 @@ void Network::ServerCommunication::sendMessage(const std::string &message) {
   std::cout << "Sent to server: " << message;
 }
 
-std::string Network::ServerCommunication::receiveMessage() {
+std::string network::ServerCommunication::receiveMessage() {
   if (!_connected || _clientFd == -1)
     throw std::runtime_error("Not connected to server");
